Scatch 백엔드 개발 #7

+++ src/main/java/NotModified304/Scatch/service/RepeatDaysService.java
@@ -1,8 +1,10 @@
 package NotModified304.Scatch.service;
 
 import NotModified304.Scatch.domain.RepeatDays;
+import NotModified304.Scatch.domain.Routine;
 import NotModified304.Scatch.dto.routine.request.RoutineCreateRequest;
 import NotModified304.Scatch.repository.interfaces.RepeatDaysRepository;
+import NotModified304.Scatch.security.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,9 +15,16 @@ import java.util.List;
 @Transactional
 @RequiredArgsConstructor
 public class RepeatDaysService {
-    private final RepeatDaysRepository repeatDaysRepository;
 
-    public void registerRepeatDays(Long routineId, List<Integer> days) {
+    private final RepeatDaysRepository repeatDaysRepository;
+    private final RoutineService routineService;
+
+    public void registerRepeatDays(String username, Long routineId, List<Integer> days) {
+
+        // 루틴 접근 권한 체크
+        Routine routine = routineService.findById(routineId);
+        SecurityUtil.validateOwner(routine.getUsername(), username);
+
         for(Integer day : days) {
             repeatDaysRepository.save(RepeatDays.builder()
                     .routineId(routineId)
@@ -25,7 +34,12 @@ public class RepeatDaysService {
     }
 
     // 반복 요일 삭제
-    public void removeRepeatDays(Long routineId) {
+    public void removeRepeatDays(String username, Long routineId) {
+
+        // 루틴 접근 권한 체크
+        Routine routine = routineService.findById(routineId);
+        SecurityUtil.validateOwner(routine.getUsername(), username);
+
         repeatDaysRepository.deleteByRoutineId(routineId);
     }
 }
