+++ src/main/java/NotModified304/Scatch/service/RoutineLogService.java
@@ -1,9 +1,11 @@
 package NotModified304.Scatch.service;
 
+import NotModified304.Scatch.domain.Routine;
 import NotModified304.Scatch.domain.RoutineLog;
 import NotModified304.Scatch.dto.routine.request.LogUpdateRequest;
 import NotModified304.Scatch.dto.routine.request.RoutineUpdateRequest;
 import NotModified304.Scatch.repository.interfaces.RoutineLogRepository;
+import NotModified304.Scatch.security.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,33 +17,46 @@ import java.util.Optional;
 @Transactional
 @RequiredArgsConstructor
 public class RoutineLogService {
+
     private final RoutineLogRepository routineLogRepository;
+    private final RoutineService routineService;
 
     // log 등록
-    public void registerRoutineLog(LogUpdateRequest dto) {
-        Optional<RoutineLog> logOpt = routineLogRepository.findLog(dto.getRoutineId(), dto.getDate());
+    public void registerRoutineLog(String username, LogUpdateRequest req) {
+
+        // 루틴 접근 권한 체크
+        Routine routine = routineService.findById(req.getRoutineId());
+        SecurityUtil.validateOwner(routine.getUsername(), username);
+
+        Optional<RoutineLog> logOpt = routineLogRepository.findLog(req.getRoutineId(), req.getDate());
 
         // 기존 로그가 있으면 달성 여부만 변경
         if(logOpt.isPresent()) {
             RoutineLog log = logOpt.get();
-            log.setIsCompleted(dto.getIsCompleted());
+            log.setIsCompleted(req.getIsCompleted());
         }
         // 기존 로그가 없으면 새로 등록
         else {
             RoutineLog log = RoutineLog.builder()
-                    .routineId(dto.getRoutineId())
-                    .date(dto.getDate())
-                    .isCompleted(dto.getIsCompleted())
+                    .routineId(req.getRoutineId())
+                    .date(req.getDate())
+                    .isCompleted(req.getIsCompleted())
                     .build();
+
             routineLogRepository.save(log);
         }
     }
 
-    // 루틴 로그 수정
-    public void updateRoutineLog(RoutineUpdateRequest dto) {
-        Long routineId = dto.getRoutineId();
-        LocalDate startDate = dto.getStartDate();
-        LocalDate endDate = dto.getEndDate();
+    // 루틴 로그 수정 (루틴 수정 시, 함께 수정됨)
+    public void updateRoutineLog(String username, RoutineUpdateRequest req) {
+
+        // 루틴 접근 권한 체크
+        Routine routine = routineService.findById(req.getRoutineId());
+        SecurityUtil.validateOwner(routine.getUsername(), username);
+
+        Long routineId = req.getRoutineId();
+        LocalDate startDate = req.getStartDate();
+        LocalDate endDate = req.getEndDate();
 
         if(startDate.isAfter(endDate)) {
             throw new IllegalArgumentException("시작일자는 종료일자보다 이전이어야 합니다.");
@@ -56,7 +71,12 @@ public class RoutineLogService {
     }
 
     // 루틴 로그 삭제
-    public void removeRoutineLog(Long routineId) {
+    public void removeRoutineLog(String username, Long routineId) {
+
+        // 루틴 접근 권한 체크
+        Routine routine = routineService.findById(routineId);
+        SecurityUtil.validateOwner(routine.getUsername(), username);
+
         routineLogRepository.deleteByRoutineId(routineId);
     }
 }
